// models/Event.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0' at 'Thu, 09 Dec 2021 17:29:41 GMT'
 */
export type Event = IContentItem<{
  dateTimeToDateTime: Elements.DateTimeElement;
  sponsors: Elements.LinkedItemsElement<IContentItem>;
  title: Elements.TextElement;
  venue: Elements.LinkedItemsElement<IContentItem>;
  description: Elements.RichTextElement;
  dateTimeFromDateTime: Elements.DateTimeElement;
  heroImage: Elements.LinkedItemsElement<IContentItem>;
  agendas: Elements.LinkedItemsElement<IContentItem>;
  journeyStage: Elements.TaxonomyElement;
}>;