// models/Sponsor.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0' at 'Thu, 09 Dec 2021 17:29:42 GMT'
 */
export type Sponsor = IContentItem<{
  media: Elements.LinkedItemsElement<IContentItem>;
  name: Elements.TextElement;
  body: Elements.RichTextElement;
}>;